Modified editing buttons into buttons

+++ src/components/MapCard/DeleteButton.tsx
@@ -4,6 +4,20 @@ import { Trash } from 'lucide-react'
 import { api } from "@/trpc/react";
 import { deleteMapImage } from '@/utils/supabaseHandler';
 
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog"
+import { Button } from '../ui/button';
+
+
 interface Props {
   mapId: number
   map_generated_name: string
@@ -12,17 +26,35 @@ interface Props {
 const DeleteButton = ({ mapId, map_generated_name }: Props) => {
   const { mutateAsync: deleteMap } = api.map.deleteMap.useMutation();
 
+  // non va magari togli il dropdown e metti le icone
+
   return (
-    <DropdownMenuItem
-      className='bg-destructive hover:bg-destructive/80 active:bg-destructive/60'
-      onClick={() => {
-        deleteMap({ id: mapId })
-        deleteMapImage(map_generated_name)
-      }}
-    >
-      Elimina
-      <DropdownMenuShortcut><Trash /></DropdownMenuShortcut>
-    </DropdownMenuItem>
+    <AlertDialog>
+      <AlertDialogTrigger asChild>
+        <Button
+          variant={"destructive"}
+        >
+          <Trash />
+        </Button>
+      </AlertDialogTrigger>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Ma te son fora?</AlertDialogTitle>
+          <AlertDialogDescription>
+            Sei sicuro di voler eliminare questa mappa?
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel>No dei</AlertDialogCancel>
+          <AlertDialogAction
+            onClick={() => {
+              deleteMap({ id: mapId })
+              deleteMapImage(map_generated_name)
+            }}
+          >Si ah</AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   )
 }
